Updated Android app to include an additional page for calculating real estate returns
