+++ app/build.gradle.kts
@@ -11,7 +11,7 @@ android {
         applicationId = "com.algorithmicworldautomation.androidinvestingcalculator"
         minSdk = 24
         targetSdk = 34
-        versionCode = 2
+        versionCode = 3
         versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
